Разбил файл со скриптами на модули
